refactor(main): extract _configure_service_startup_stages to backend/core/registrations/core_startup (audit A8, refs #153)
Commit 4 of N in PR A8.

Same mechanical pattern + two corrections:
- core_startup.configure() imports + calls the other 3 modules
  directly (group2_repositories, group2_services, phase4) instead
  of bare-name calls.
- group2_services.py needed _create_database_engine helper (lifted
  from main.py) + EntityService import.

  backend/core/registrations/core_startup.py   (new, ~388 LOC)
  backend/core/registrations/group2_services.py (+_create_database_engine helper, +imports)
  backend/main.py                               (-330 LOC: 1603 -> 1272)

Tests: 828 passed, 0 failed, 28 skipped (matches main pre-extraction).

Refs #145, refs #153
